Move auth schema to schema directory and add auth types to types directory
