--- web/src/auth.schema.ts
@@ -1,41 +1,0 @@
-import {z} from "zod";
-//Schema for Login
- export const loginSchema = z.object({
-  username: z.string().min(1, "Username is required"),
-  password: z.string().min(1, "Password is required"),
-});
-
-
-
-//Schema for Register
-export const registerSchema = z.object({
-  username: z
-  .string()
-  .trim()
-  .min(3, "Username must be at least 3 characters")
-  .max(20, "Username must be at most 20 characters")
-  .regex(/^[a-zA-Z0-9_]+$/, "Username can only contain letters, numbers, and underscores"),
-
-
-email:z.email({message:"Invalid email address"}),  
-
-password: z
-  .string()
-  .trim()
-  .min(8, "Password must be at least 8 characters long")
-  .regex(/[A-Z]/, "Password must contain at least one uppercase letter")
-  .regex(/[a-z]/, "Password must contain at least one lowercase letter")
-  .regex(/[0-9]/, "Password must contain at least one number")
-  .regex(/[@$!%*?&]/, "Password must contain at least one special character (@, $, !, %, *, ?, &)")
-  .regex(/^\S+$/, "Password cannot contain spaces"),
-
-  confirmPassword: z
-    .string()
-    .trim(),
-})
-.refine((data) => data.password === data.confirmPassword, {
-  message: "Passwords do not match",
-  path: ["confirmPassword"],
-});
-
-
